mavericks-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
PyPDF2==3.0.1
cohere>=4.0.0